+++ client/src/ui/date-picker/modular-forms-date-range-picker.tsx
@@ -4,7 +4,7 @@ import {
 	type FieldValues,
 	type FormStore,
 } from "@modular-forms/solid";
-import { spreadProps } from "core/utils/utils";
+import { spreadProps } from "core/utils";
 import { partial } from "es-toolkit";
 import {
 	DateRangePicker,
