+++ client/src/ui/hover-card/index.tsx
@@ -3,7 +3,7 @@ import {
 	HoverCard as HoverCardPrimitive,
 } from "@kobalte/core/hover-card";
 import type { PolymorphicProps } from "@kobalte/core/polymorphic";
-import { spreadProps } from "core/utils/utils";
+import { spreadProps } from "core/utils";
 import type { ValidComponent } from "solid-js";
 import { cn } from "ui/utils";
