+++ client/src/ui/progress/index.tsx
@@ -3,7 +3,7 @@ import {
 	type ProgressRootProps,
 	Progress as ProgressPrimitive,
 } from "@kobalte/core/progress";
-import { spreadProps } from "core/utils/utils";
+import { spreadProps } from "core/utils";
 import type { ValidComponent } from "solid-js";
 import { cn } from "ui/utils";
