+++ client/src/ui/separator/index.tsx
@@ -3,7 +3,7 @@ import {
 	Separator as SeparatorPrimitive,
 	type SeparatorRootProps,
 } from "@kobalte/core/separator";
-import { spreadProps } from "core/utils/utils";
+import { spreadProps } from "core/utils";
 import type { ValidComponent } from "solid-js";
 import { cn } from "ui/utils";
